- move around some parts of the code
