+++ src/configuration.ts
@@ -1,6 +1,7 @@
 import {Extractor, Regex, Rule, Sort, Transformer} from 'github-pr-collector/lib/configuration'
+import {PullConfiguration} from 'github-pr-collector/lib/types'
 
-export interface Configuration {
+export interface Configuration extends PullConfiguration {
   max_tags_to_fetch: number
   max_pull_requests: number
   max_back_track_time_days: number
